refactor(status): extract module to reduce top-level item count (#4965)
